updated readme with usage examples; moved test fixture; updated copyright in license statement
