+++ COPYING
@@ -1,5 +1,6 @@
 WeighWords: a Python library for creating word weights/word clouds from text
-Copyright 2011 University of Amsterdam
+Copyright 2011-2019 University of Amsterdam
+Copyright 2019 TinQwise Stamkracht
 
 This program is free software: you can redistribute it and/or modify it under
 the terms of the GNU Lesser General Public License as published by the
